Conjure Distortion - lv3 Conj/Tloc for Warpers
Creates a cloud of distortion in a random adjacent empty (no monster AND no cloud of distortion) location.
Clouds of distortion produce all distortion effects except for banishment - damaging, blinking and teleporting any being standing in them.

In the Warper starting book.

The utility of the cloud is comparable to Conjure Flame but more random - rather than steady damage that always kills an ogre in 2-3 hits, you may get a tele/blink faster, a bad blink that makes the situation worse or you might get lucky damage.

While it is much harder to position than Conjure Flame, Warper has the tools necessary to deal with this - namely, Blink, Lesser Beckoning and Gell's Gravitas. Still, due to its inherent randomness, it is not as reliable a damage/escape tool as Conjure Flame is, and this is by design.

You can also use Conjure Distortion and deliberately jump into it, hoping to get the 10% chance of an instant teleport or 10% chance of a delayed teleport, but you may just get blinked out of it or repeatedly damaged instead.

Banish is left out of the effects list because giving Warpers random banish is dumb.