Rename crawl's custom `require`
We don't (normally) load the lua `package` lib, but this function
shadows a lua `require` that would be loaded by this lib. Rename our
`require` both to make it clear that this *isn't* the real lua require,
and to allow the real thing to work (especially when libraries try to
use it) for builds where `package` is loaded.