Fix cross-platform lua number->int conversion the hard way
Lua 5.1 doesn't distinguish integers from numbers generally, and stores
numbers as doubles, but on the c++ side we very often want ints. In
3d66cf40a65a384 I discovered that, depending on where the lua library is
compiled, it may use different conversion methods in its conversion
calls, which have different truncation behavior, leading to seed
instability across systems (where it uses a specific hack on 32bit
non-sse builds, a different one in msvc, and a regular cast otherwise).
This behavior is determined at compile-time in luaconf.h, and the first
attempt at a fix tried to prevent contrib lua from being built with the
alternate behavior, always using a cast. This didn't generalize for at
least two reasons: it creates problems trying to compile on other archs
(see #992, #996), and it doesn't even cover all the problem situations:
linux will typically build with system lua, and then we are at the mercy
of whatever arch the system lua was built on anyways. (I have replicated
the problem in a 32 bit ubuntu container, for example, even with the
compilation fix.)

This commit moves away from using lua's default calls for converting to
ints, to conversion functions that we control in crawlcode. It's overkill
to convert everything, only the dungeon and related calls are really
needed, but converting everything was the most robust way to do it. Note
that this *doesn't* address any instances of this call in the lua lib,
with the hope that they won't matter.

The only alternative that I can think of to a change like this would be
to require building with contrib lua, and then fix the problem there,
but this would seem to go against the spirit of our current build
system.