Throwing Overhaul III: Remove most tomahawk and javelin brands
Now that blowguns are removed, throwing users do not need three ways to
poison a monster.

Steel is removed; silver now gets both bonuses, to be a "rare extra
damage throwable" that fits a niche not available with other weapon
types. This might be too good, which can be addressed in a couple of
ways.

Exploding, while "neat" was rarely different from yet another bonus
damage throwable.

Returning and penetration are removed, these effects will return in a
different way in subsequent commits.

[Committer's note: Closes #1048. Significant expansion of the original
 patch]