Rename frenzy and blinding darts
This commit is separate as renames tend to be controversial. All changes
are cosmetic.

Darts of frenzy become "datura-tipped darts" after the genus of toxic
plant that can cause psychosis and darts of blinding become
"atropa-tipped darts" after the genus of toxic plant that can cause
visual distortion and confusion.

Players aren't likely to know this (unless they're a botanist), but they
probably also don't know that curare is (in addition to being a generic
name for blowgun poisons) a genus of toxic plant. So it fits the
"exotic" dart poisons into a nice family.